added folder for qa_dataset, fixed roberta jenkins, added docstring
