+++ examples/nlp/question_answering/conf/config.yaml
@@ -19,6 +19,7 @@ trainer:
 
 model:
 
+  checkpoint: null # model checkpoint path to restore weights 
   version_2_with_negative: false
   # If true, the examples contain some that do not have an answer.
