+++ examples/nlp/question_answering/question_answering_squad.py
@@ -28,7 +28,7 @@ def main(cfg: DictConfig) -> None:
     trainer = pl.Trainer(**cfg.trainer)
     exp_manager(trainer, cfg.get("exp_manager", None))
     question_answering_model = QAModel(cfg.model, trainer=trainer)
-    question_answering_model.restore_from(cfg.model.language_model.bert_checkpoint)
+    question_answering_model.restore_from(cfg.model.checkpoint)
     trainer.fit(question_answering_model)
