+++ nemo/collections/nlp/data/__init__.py
@@ -14,6 +14,7 @@
 
 from nemo.collections.nlp.data.data_utils import *
 from nemo.collections.nlp.data.lm_bert_dataset import *
+from nemo.collections.nlp.data.question_answering_squad.qa_dataset import SquadDataset
 from nemo.collections.nlp.data.token_classification.token_classification_dataset import (
     BertTokenClassificationDataset,
     BertTokenClassificationInferDataset,
